Refactoring homeController
